# Conflicts:
#	src/main/java/poe/command/EvolveCharacter.java
#	src/main/java/poe/command/model/EvolutionStatus.java
#	src/main/java/poe/evaluator/spring/SpringCharacterEvaluator.java
#	src/main/java/poe/jenetics/EvolutionWatcher.java
#	src/main/java/poe/jenetics/FitnessFunction.java

+++ src/main/java/poe/jenetics/DeterministicMutator.java
@@ -5,7 +5,7 @@ import java.util.stream.IntStream;
 import org.jenetics.Mutator;
 import org.jenetics.util.MSeq;
 
-public class DeterministicMutator extends Mutator<SkillGene, Integer>
+public class DeterministicMutator extends Mutator<SkillGene, Float>
 {
 	private final List<List<Integer>> testingGenes;
